fix compiler warning
compiler for 8266 complained about "comparing singed to unsigned"

+++ wled00/FX_fcn.cpp
@@ -1276,7 +1276,7 @@ void WS2812FX::service() {
 
   _isServicing = true;
   bool doShow = false;    // true if ≥1 active segment was processed (and strip was not suspended mid-loop) → triggers show()
-  for (int i = 0; i < _segments.size(); i++) {
+  for (size_t i = 0; i < _segments.size(); i++) {
     Segment &seg = _segments[i];
    _segment_index = i;
     if (_suspend) break; // abort processing segments if suspend requested during service()
